update to DIGITS3.3 and add pad_fill

+++ digits/views.py
@@ -1,26 +1,27 @@
-# Copyright (c) 2014-2015, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2014-2016, NVIDIA CORPORATION.  All rights reserved.
+from __future__ import absolute_import
 
-import json
-import traceback
 import glob
+import json
 import platform
+import traceback
 
 import flask
+from flask.ext.socketio import join_room, leave_room
 from werkzeug import HTTP_STATUS_CODES
 import werkzeug.exceptions
-from flask.ext.socketio import join_room, leave_room
 
-from . import dataset, model
-from config import config_value
-from webapp import app, socketio, scheduler, autodoc
-import dataset.views
-import model.views
-from digits.utils import errors
+from .config import config_value
+from .webapp import app, socketio, scheduler
+import digits
+from digits import dataset, model, utils
+from digits.log import logger
 from digits.utils.routing import request_wants_json
 
-@app.route('/index.json', methods=['GET'])
-@app.route('/', methods=['GET'])
-@autodoc(['home', 'api'])
+blueprint = flask.Blueprint(__name__, __name__)
+
+@blueprint.route('/index.json', methods=['GET'])
+@blueprint.route('/', methods=['GET'])
 def home():
     """
     DIGITS home page
@@ -38,24 +39,29 @@ def home():
     completed_models    = get_job_list(model.ModelJob, False)
 
     if request_wants_json():
-        return flask.jsonify({
-            'datasets': [j.json_dict()
-                for j in running_datasets + completed_datasets],
-            'models': [j.json_dict()
-                for j in running_models + completed_models],
-            })
+        data = {
+                'version': digits.__version__,
+                'jobs_dir': config_value('jobs_dir'),
+                'datasets': [j.json_dict()
+                    for j in running_datasets + completed_datasets],
+                'models': [j.json_dict()
+                    for j in running_models + completed_models],
+                }
+        if config_value('server_name'):
+            data['server_name'] = config_value('server_name')
+        return flask.jsonify(data)
     else:
         new_dataset_options = [
                 ('Images', [
                     {
                         'title': 'Classification',
                         'id': 'image-classification',
-                        'url': flask.url_for('image_classification_dataset_new'),
+                        'url': flask.url_for('digits.dataset.images.classification.views.new'),
                         },
                     {
                         'title': 'Other',
                         'id': 'image-generic',
-                        'url': flask.url_for('generic_image_dataset_new'),
+                        'url': flask.url_for('digits.dataset.images.generic.views.new'),
                         },
                     ])
                 ]
@@ -64,12 +70,12 @@ def home():
                     {
                         'title': 'Classification',
                         'id': 'image-classification',
-                        'url': flask.url_for('image_classification_model_new'),
+                        'url': flask.url_for('digits.model.images.classification.views.new'),
                         },
                     {
                         'title': 'Other',
                         'id': 'image-generic',
-                        'url': flask.url_for('generic_image_model_new'),
+                        'url': flask.url_for('digits.model.images.generic.views.new'),
                         },
                     ])
                 ]
@@ -87,16 +93,57 @@ def home():
 
 def get_job_list(cls, running):
     return sorted(
-            [j for j in scheduler.jobs if isinstance(j, cls) and j.status.is_running() == running],
+            [j for j in scheduler.jobs.values() if isinstance(j, cls) and j.status.is_running() == running],
             key=lambda j: j.status_history[0][1],
             reverse=True,
             )
 
 
+### Authentication/login
+
+@blueprint.route('/login', methods=['GET','POST'])
+def login():
+    """
+    Ask for a username (no password required)
+    Sets a cookie
+    """
+    # Get the URL to redirect to after logging in
+    next_url = utils.routing.get_request_arg('next') or \
+            flask.request.referrer or flask.url_for('.home')
+
+    if flask.request.method == 'GET':
+        return flask.render_template('login.html', next=next_url)
+
+    # Validate username
+    username = utils.routing.get_request_arg('username').strip()
+    try:
+        utils.auth.validate_username(username)
+    except ValueError as e:
+        # Invalid username
+        flask.flash(e.message, 'danger')
+        return flask.render_template('login.html', next=next_url)
+
+    # Valid username
+    response = flask.make_response(flask.redirect(next_url))
+    response.set_cookie('username', username)
+    return response
+
+@blueprint.route('/logout', methods=['GET','POST'])
+def logout():
+    """
+    Unset the username cookie
+    """
+    next_url = utils.routing.get_request_arg('next') or \
+            flask.request.referrer or flask.url_for('.home')
+
+    response = flask.make_response(flask.redirect(next_url))
+    response.set_cookie('username', '', expires=0)
+    return response
+
+
 ### Jobs routes
 
-@app.route('/jobs/<job_id>', methods=['GET'])
-@autodoc('jobs')
+@blueprint.route('/jobs/<job_id>', methods=['GET'])
 def show_job(job_id):
     """
     Redirects to the appropriate /datasets/ or /models/ page
@@ -106,30 +153,46 @@ def show_job(job_id):
         raise werkzeug.exceptions.NotFound('Job not found')
 
     if isinstance(job, dataset.DatasetJob):
-        return flask.redirect(flask.url_for('datasets_show', job_id=job_id))
+        return flask.redirect(flask.url_for('digits.dataset.views.show', job_id=job_id))
     if isinstance(job, model.ModelJob):
-        return flask.redirect(flask.url_for('models_show', job_id=job_id))
+        return flask.redirect(flask.url_for('digits.model.views.show', job_id=job_id))
     else:
         raise werkzeug.exceptions.BadRequest('Invalid job type')
 
-@app.route('/jobs/<job_id>', methods=['PUT'])
-@autodoc('jobs')
+@blueprint.route('/jobs/<job_id>', methods=['PUT'])
+@utils.auth.requires_login(redirect=False)
 def edit_job(job_id):
     """
-    Edit the name of a job
+    Edit a job's name and/or notes
     """
     job = scheduler.get_job(job_id)
     if job is None:
         raise werkzeug.exceptions.NotFound('Job not found')
 
-    old_name = job.name()
-    job._name = flask.request.form['job_name']
-    return 'Changed job name from "%s" to "%s"' % (old_name, job.name())
+    if not utils.auth.has_permission(job, 'edit'):
+        raise werkzeug.exceptions.Forbidden()
 
-@app.route('/datasets/<job_id>/status', methods=['GET'])
-@app.route('/models/<job_id>/status', methods=['GET'])
-@app.route('/jobs/<job_id>/status', methods=['GET'])
-@autodoc('jobs')
+    # Edit name
+    if 'job_name' in flask.request.form:
+        name = flask.request.form['job_name'].strip()
+        if not name:
+            raise werkzeug.exceptions.BadRequest('name cannot be blank')
+        job._name = name
+        logger.info('Set name to "%s".' % job.name(), job_id=job.id())
+
+    # Edit notes
+    if 'job_notes' in flask.request.form:
+        notes = flask.request.form['job_notes'].strip()
+        if not notes:
+            notes = None
+        job._notes = notes
+        logger.info('Updated notes.', job_id=job.id())
+
+    return '%s updated.' % job.job_type()
+
+@blueprint.route('/datasets/<job_id>/status', methods=['GET'])
+@blueprint.route('/models/<job_id>/status', methods=['GET'])
+@blueprint.route('/jobs/<job_id>/status', methods=['GET'])
 def job_status(job_id):
     """
     Returns a JSON objecting representing the status of a job
@@ -145,10 +208,10 @@ def job_status(job_id):
         result['type'] = job.job_type()
     return json.dumps(result)
 
-@app.route('/datasets/<job_id>', methods=['DELETE'])
-@app.route('/models/<job_id>', methods=['DELETE'])
-@app.route('/jobs/<job_id>', methods=['DELETE'])
-@autodoc('jobs')
+@blueprint.route('/datasets/<job_id>', methods=['DELETE'])
+@blueprint.route('/models/<job_id>', methods=['DELETE'])
+@blueprint.route('/jobs/<job_id>', methods=['DELETE'])
+@utils.auth.requires_login(redirect=False)
 def delete_job(job_id):
     """
     Deletes a job
@@ -157,18 +220,21 @@ def delete_job(job_id):
     if job is None:
         raise werkzeug.exceptions.NotFound('Job not found')
 
+    if not utils.auth.has_permission(job, 'delete'):
+        raise werkzeug.exceptions.Forbidden()
+
     try:
         if scheduler.delete_job(job_id):
             return 'Job deleted.'
         else:
             raise werkzeug.exceptions.Forbidden('Job not deleted')
-    except errors.DeleteError as e:
+    except utils.errors.DeleteError as e:
         raise werkzeug.exceptions.Forbidden(str(e))
 
-@app.route('/datasets/<job_id>/abort', methods=['POST'])
-@app.route('/models/<job_id>/abort', methods=['POST'])
-@app.route('/jobs/<job_id>/abort', methods=['POST'])
-@autodoc('jobs')
+@blueprint.route('/datasets/<job_id>/abort', methods=['POST'])
+@blueprint.route('/models/<job_id>/abort', methods=['POST'])
+@blueprint.route('/jobs/<job_id>/abort', methods=['POST'])
+@utils.auth.requires_login(redirect=False)
 def abort_job(job_id):
     """
     Aborts a running job
@@ -181,6 +247,30 @@ def abort_job(job_id):
         return 'Job aborted.'
     else:
         raise werkzeug.exceptions.Forbidden('Job not aborted')
+
+@blueprint.route('/clone/<clone>', methods=['POST', 'GET'])
+@utils.auth.requires_login
+def clone_job(clone):
+    """
+    Clones a job with the id <clone>, populating the creation page with data saved in <clone>
+    """
+
+    ## <clone> is the job_id to clone
+
+    job = scheduler.get_job(clone)
+    if job is None:
+        raise werkzeug.exceptions.NotFound('Job not found')
+
+    if isinstance(job, dataset.ImageClassificationDatasetJob):
+        return flask.redirect(flask.url_for('digits.dataset.images.classification.views.new') + '?clone=' + clone)
+    if isinstance(job, dataset.GenericImageDatasetJob):
+        return flask.redirect(flask.url_for('digits.dataset.images.generic.views.new') + '?clone=' + clone)
+    if isinstance(job, model.ImageClassificationModelJob):
+        return flask.redirect(flask.url_for('digits.model.images.classification.views.new') + '?clone=' + clone)
+    if isinstance(job, model.GenericImageModelJob):
+        return flask.redirect(flask.url_for('digits.model.images.generic.views.new') + '?clone=' + clone)
+    else:
+        raise werkzeug.exceptions.BadRequest('Invalid job type')
 
 ### Error handling
 
@@ -226,8 +316,7 @@ for code in HTTP_STATUS_CODES:
 
 ### File serving
 
-@app.route('/files/<path:path>', methods=['GET'])
-@autodoc('util')
+@blueprint.route('/files/<path:path>', methods=['GET'])
 def serve_file(path):
     """
     Return a file in the jobs directory
@@ -240,8 +329,7 @@ def serve_file(path):
 
 ### Path Completion
 
-@app.route('/autocomplete/path', methods=['GET'])
-@autodoc('util')
+@blueprint.route('/autocomplete/path', methods=['GET'])
 def path_autocomplete():
     """
     Return a list of paths matching the specified preamble
@@ -254,7 +342,7 @@ def path_autocomplete():
         suggestions = [p.replace('\\', '/') for p in suggestions]
 
     result = {
-        "suggestions": suggestions
+        "suggestions": sorted(suggestions)
     }
 
     return json.dumps(result)
